requirements.txt
cffi==1.15.1
distro==1.8.0
numpy==1.24.1
packaging==22.0
Pillow==9.4.0
pillow-heif==0.9.1
pycparser==2.21
scikit-build==0.16.4
tkinterdnd2==0.3.0
whatimage==0.0.3
